refactor EditableTextares
